Back out anticipated epoch changes
After all the debate as to what to do on maint, we are going to end up just
incorporating epoch into the version string, so we don't need this separate
field at all. Revert commit 5c8083baa4a and also kill the force flag we were
recording here as well.

Signed-off-by: Dan McGee <dan@archlinux.org>